refactor: reorganize runner and risk modules
