+++ backend/tests/test_regime_detector.py
@@ -1,6 +1,6 @@
 import unittest
 
-from analysis.regime_detector import RegimeDetector
+from piphawk_ai.analysis.regime_detector import RegimeDetector
 
 
 class DummyTick(dict):
