+++ piphawk_ai/__init__.py
@@ -1,4 +1,5 @@
 """Namespace package for piphawk AI."""
+from pkgutil import extend_path
 import importlib
 import sys
 
@@ -13,9 +14,12 @@ _submodules = [
     "strategies",
     "regime",
     "core",
+    "runner",
     "models",
     "training",
 ]
+__path__ = extend_path(__path__, __name__)
+
 for _name in _submodules:
     try:
         module = importlib.import_module(_name)
