+++ pyproject.toml
@@ -3,9 +3,53 @@ requires = ["setuptools>=61"]
 build-backend = "setuptools.build_meta"
 
 [project]
-name = "piphawk_ai"
-version = "0.5.0"
-requires-python = ">=3.11"
+name = "piphawk-ai"
+version = "0.5.0.dev0"
+requires-python = ">=3.10"
+dependencies = [
+    "annotated-types==0.7.0",
+    "anyio==4.9.0",
+    "certifi==2025.4.26",
+    "charset-normalizer==3.4.2",
+    "click==8.2.0",
+    "distro==1.9.0",
+    "exceptiongroup==1.3.0",
+    "fastapi==0.115.12",
+    "h11==0.16.0",
+    "httpcore==1.0.9",
+    "httpx==0.28.1",
+    "idna==3.10",
+    "jiter==0.9.0",
+    "numpy==2.2.5",
+    "openai==1.78.0",
+    "pandas==2.2.3",
+    "pydantic==2.11.4",
+    "pydantic_core==2.33.2",
+    "python-dateutil==2.9.0.post0",
+    "python-dotenv==1.1.0",
+    "pytz==2025.2",
+    "requests==2.32.3",
+    "six==1.17.0",
+    "sniffio==1.3.1",
+    "starlette==0.46.2",
+    "tqdm==4.67.1",
+    "typing-inspection==0.4.0",
+    "typing_extensions==4.13.2",
+    "tzdata==2025.2",
+    "urllib3==2.4.0",
+    "uvicorn==0.34.2",
+    "docker==7.0.0",
+    "line-bot-sdk==3.17.1",
+    "apscheduler==3.10.4",
+    "PyYAML==6.0.1",
+    "scikit-learn==1.4.2",
+    "mabwiser==2.7.4",
+    "kafka-python==2.0.2",
+    "prometheus-client==0.20.0",
+    "hdbscan==0.8.33",
+    "d3rlpy==2.1.0",
+    "transformers==4.41.2",
+]
 
 [tool.setuptools]
 packages = ["piphawk_ai"]
