+++ risk/__init__.py
@@ -1,5 +1,5 @@
 from .trade_guard import TradeGuard
 from .cvar import calc_cvar
-from .portfolio_risk_manager import PortfolioRiskManager
+from piphawk_ai.risk.manager import PortfolioRiskManager
 
 __all__ = ["TradeGuard", "calc_cvar", "PortfolioRiskManager"]
